requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
locust
tqdm

# Setup modules
gevent
flask

# default available packages for python3action
# beautifulsoup4==4.8.0
# httplib2==0.13.0
# kafka_python==1.4.6
# lxml==4.3.4
# python-dateutil==2.8.0
# requests==2.22.0
# scrapy==1.6.0
# simplejson==3.16.0
# virtualenv==16.7.1
# twisted==20.3.0
# PyJWT==1.7.1

# custom packages
minio
Pillow
pymongo
redis
cloudant

aiohttp
